requirements.txt
aniso8601==8.0.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
